requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
torchmetrics[all]
pesq
pyctcdecode
soundfile
numpy
tqdm
matplotlib
pandas
wandb
hydra_core

black
isort
pre-commit
